+++ deploy/dynamo/sdk/src/dynamo/sdk/cli/utils.py
@@ -23,10 +23,15 @@ import pathlib
 import random
 import socket
 import typing as t
+import collections
+import yaml
 
 import click
 import psutil
 from click import Command, Context
+
+from dynamo.sdk.lib.logging import configure_server_logging
+configure_server_logging()
 
 logger = logging.getLogger(__name__)
 
@@ -182,3 +187,146 @@ def save_dynamo_state(
         json.dump(state, f)
 
     logger.warning(f"Saved state to {state_file}")
+
+
+def _parse_service_arg(arg_name: str, arg_value: str) -> tuple[str, str, t.Any]:
+    """Parse a single CLI argument into service name, key, and value."""
+
+    parts = arg_name.split(".")
+    service = parts[0]
+    nested_keys = parts[1:]
+
+    # Special case: if this is a ServiceArgs.envs.* path, keep value as string
+    if (
+        len(nested_keys) >= 2
+        and nested_keys[0] == "ServiceArgs"
+        and nested_keys[1] == "envs"
+    ):
+        value: t.Union[str, int, float, bool, dict, list] = arg_value
+    else:
+        # Parse value based on type for non-env vars
+        try:
+            value = json.loads(arg_value)
+        except json.JSONDecodeError:
+            if arg_value.isdigit():
+                value = int(arg_value)
+            elif arg_value.replace(".", "", 1).isdigit() and arg_value.count(".") <= 1:
+                value = float(arg_value)
+            elif arg_value.lower() in ("true", "false"):
+                value = arg_value.lower() == "true"
+            else:
+                value = arg_value
+
+    # Build nested dict structure
+    result = value
+    for key in reversed(nested_keys[1:]):
+        result = {key: result}
+
+    return service, nested_keys[0], result
+
+
+def _parse_service_args(args: list[str]) -> t.Dict[str, t.Any]:
+    service_configs: t.DefaultDict[str, t.Dict[str, t.Any]] = collections.defaultdict(
+        dict
+    )
+
+    def deep_update(d: dict, key: str, value: t.Any):
+        """
+        Recursively updates nested dictionaries. We use this to process arguments like
+
+        ---Worker.ServiceArgs.env.CUDA_VISIBLE_DEVICES="0,1"
+
+        The _parse_service_arg function will parse this into:
+        service = "Worker"
+        nested_keys = ["ServiceArgs", "envs", "CUDA_VISIBLE_DEVICES"]
+
+        And returns returns: ("VllmWorker", "ServiceArgs", {"envs": {"CUDA_VISIBLE_DEVICES": "0,1"}})
+
+        We then use deep_update to update the service_configs dictionary with this nested value.
+        """
+        if isinstance(value, dict) and key in d and isinstance(d[key], dict):
+            for k, v in value.items():
+                deep_update(d[key], k, v)
+        else:
+            d[key] = value
+
+    index = 0
+    while index < len(args):
+        next_arg = args[index]
+
+        if not (next_arg.startswith("--") or "." not in next_arg):
+            continue
+        try:
+            if "=" in next_arg:
+                arg_name, arg_value = next_arg.split("=", 1)
+                index += 1
+            elif args[index + 1] == "=":
+                arg_name = next_arg
+                arg_value = args[index + 2]
+                index += 3
+            else:
+                arg_name = next_arg
+                arg_value = args[index + 1]
+                index += 2
+            if arg_value.startswith("-"):
+                raise ValueError("Service arg value can not start with -")
+            arg_name = arg_name[2:]
+            service, key, value = _parse_service_arg(arg_name, arg_value)
+            deep_update(service_configs[service], key, value)
+        except Exception:
+            raise ValueError(f"Error parsing service arg: {args[index]}")
+
+    return service_configs
+
+
+def resolve_service_config(
+    config_file: str | t.TextIO | None = None,
+    args: list[str] | None = None,
+) -> dict[str, dict[str, t.Any]]:
+    """Resolve service configuration from file and command line arguments.
+    
+    Args:
+        config_file: Path to YAML config file or file object
+        args: List of command line arguments
+        
+    Returns:
+        Dictionary mapping service names to their configurations
+    """
+    service_configs: dict[str, dict[str, t.Any]] = {}
+
+    # Check for deployment config first
+    if "DYN_DEPLOYMENT_CONFIG" in os.environ:
+        try:
+            deployment_config = yaml.safe_load(os.environ["DYN_DEPLOYMENT_CONFIG"])
+            # Use deployment config directly
+            service_configs = deployment_config
+            logger.info(f"Successfully loaded deployment config: {service_configs}")
+            logger.warning("DYN_DEPLOYMENT_CONFIG found in environment - ignoring configuration file and command line arguments")
+        except Exception as e:
+            logger.warning(f"Failed to parse DYN_DEPLOYMENT_CONFIG: {e}")
+    else:
+        # Load file if provided
+        if config_file:
+            with open(config_file) if isinstance(config_file, str) else contextlib.nullcontext(config_file) as f:
+                yaml_configs = yaml.safe_load(f)
+                logger.debug(f"Loaded config from file: {yaml_configs}")
+                # Initialize service_configs as empty dict if it's None
+                # Convert nested YAML structure to flat dict with dot notation
+                for service, configs in yaml_configs.items():
+                    if service not in service_configs:
+                        service_configs[service] = {}
+                    for key, value in configs.items():
+                        service_configs[service][key] = value
+
+        # Process service-specific options
+        if args:
+            cmdline_overrides = _parse_service_args(args)
+            logger.debug(f"Applying command line overrides: {cmdline_overrides}")
+            for service, configs in cmdline_overrides.items():
+                if service not in service_configs:
+                    service_configs[service] = {}
+                for key, value in configs.items():
+                    service_configs[service][key] = value
+
+    logger.debug(f"Final resolved config: {service_configs}")
+    return service_configs
