+++ examples/hello_world/config.yaml
@@ -1,0 +1,36 @@
+# SPDX-FileCopyrightText: Copyright (c) 2025 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+# SPDX-License-Identifier: Apache-2.0
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+Frontend:
+  message: "Hello from Frontend"
+  port: 8000
+  ServiceArgs:
+    workers: 1
+    resources:
+      cpu: "1"
+
+Middle:
+  message: "Hello from Middle"
+  ServiceArgs:
+    workers: 1
+    resources:
+      cpu: "1"
+
+Backend:
+  message: "Hello from Backend"
+  ServiceArgs:
+    workers: 2
+    resources:
+      cpu: "1"
